유지보수를 위한 SNSBox 추상화 계층 구분
