+++ containers/Contact/Box.tsx
@@ -1,30 +1,15 @@
 import styled from 'styled-components';
 
 import { media } from '../../styles/theme';
-import { Gmail, Kakao, DPM } from '../../public';
+import { DPM } from '../../public';
 import { Device } from '../../contexts/device';
+import SNSBox from './SNSBox';
 
 type Props = {
   device: Device;
 };
 
 export default function Box({ device }: Props) {
-  const snsList = [
-    {
-      name: 'gmail',
-      content: 'depromeet@gmail.com',
-      logo: <Gmail />,
-      onClick: () =>
-        window.open('mailto:depromeet@gmail.com?subject=[홈페이지 문의] '),
-    },
-    {
-      name: 'kakao plus friend',
-      content: 'depromeet',
-      logo: <Kakao />,
-      onClick: () => window.open('http://pf.kakao.com/_xoxmcxed/chat'),
-    },
-  ];
-
   return (
     <Container>
       {device !== 'desktop' && (
@@ -40,15 +25,7 @@ export default function Box({ device }: Props) {
         <p>디자인과 개발 관련된 일상적인 대화도</p>
         <p>언제든지 환영합니다.</p>
       </SubTitle>
-      <SNS>
-        {snsList.map(({ name, content, logo, onClick }) => (
-          <SNSButton onClick={onClick} key={name}>
-            <SNSLogo>{logo}</SNSLogo>
-            <SNSName>{name}</SNSName>
-            <SNSContent>{content}</SNSContent>
-          </SNSButton>
-        ))}
-      </SNS>
+      <SNSBox />
     </Container>
   );
 }
@@ -97,79 +74,4 @@ const SubTitle = styled.div`
     line-height: 24px;
     text-align: center;
   }
-`;
-
-const SNS = styled.div`
-  border-radius: 20px;
-  padding: 30px 20px;
-  margin-top: 20px;
-
-  backdrop-filter: blur(120px);
-  background: linear-gradient(
-    126.6deg,
-    rgba(255, 255, 255, 0.16) 28.69%,
-    rgba(255, 255, 255, 0) 100%
-  );
-
-  ${media.mobile} {
-    margin-top: 20px;
-    padding: 30px 20px;
-    border-radius: 20px;
-  }
-`;
-
-const SNSButton = styled.button`
-  position: relative;
-
-  display: flex;
-  align-items: center;
-
-  width: 100%;
-  height: 40px;
-  border-radius: 31px;
-
-  color: #000;
-  border: none;
-  background-color: #fff;
-
-  ${media.mobile} {
-    box-sizing: border-box;
-    padding: 1.2rem 2rem;
-  }
-
-  :hover {
-    background-color: #c0c0c0;
-  }
-
-  & + & {
-    margin-top: 20px;
-  }
-`;
-
-const SNSLogo = styled.div`
-  margin-left: 1.6rem;
-
-  ${media.mobile} {
-    margin-left: 0;
-  }
-`;
-
-const SNSName = styled.div`
-  font-size: 1.2rem;
-  font-weight: 700;
-  margin-left: 0.915rem;
-
-  ${media.mobile} {
-    font-weight: bold;
-    line-height: 1.5rem;
-  }
-`;
-
-const SNSContent = styled.div`
-  position: absolute;
-  right: 0;
-
-  margin-right: 2.4rem;
-  font-size: 1.4rem;
-  font-weight: 500;
 `;
